requirements.txt
python-telegram-bot == 20.3
Biopython==1.81
requests~=2.30.0
python-dotenv~=1.0.0